pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.6.1",
    "werkzeug>=3.1.3",
    "yfinance>=0.2.59",
    "nltk>=3.9.1",
    "numpy>=2.2.5",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "python-dotenv>=1.1.0",
    "dnspython>=2.7.0",
    "pymongo>=4.12.1",
]
